:recycle: Moved events under scm
