+++ lib/incomingHandler.js
@@ -1,11 +1,11 @@
 "use strict";
 
 // Event handlers
-const checkSuiteEvent = require("../events/checkSuite");
-const checkRunEvent = require("../events/checkRun");
-const pullRequestEvent = require("../events/pullRequest");
-const pushEvent = require("../events/push");
-const releaseEvent = require("../events/release");
+const checkSuiteEvent = require("../scm/events/checkSuite");
+const checkRunEvent = require("../scm/events/checkRun");
+const pullRequestEvent = require("../scm/events/pullRequest");
+const pushEvent = require("../scm/events/push");
+const releaseEvent = require("../scm/events/release");
 
 /**
  * handle task update
