+++ scm/events/pullRequest.js
@@ -1,9 +1,9 @@
 "use strict";
 
-const checkRun = require("../lib/checkRun");
-const notification = require("../lib/notification");
-const config = require("../lib/config");
-const build = require("../lib/build");
+const checkRun = require("../../lib/checkRun");
+const notification = require("../../lib/notification");
+const config = require("../../lib/config");
+const build = require("../../lib/build");
 
 /**
  * handle event
